backend/requirements.txt
asgiref==3.5.2
Django==3.2.14
django-colorfield==0.7.2
django-extra-fields==3.0.2
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
flake8==5.0.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
